create_tool_from_function + decorator
