+++ docs/pydoc/config/tools_api.yml
@@ -2,7 +2,7 @@ loaders:
   - type: haystack_pydoc_tools.loaders.CustomPythonLoader
     search_path: [../../../haystack/tools]
     modules:
-      ["tool"]
+      ["tool", "from_function"]
     ignore_when_discovered: ["__init__"]
 processors:
   - type: filter
